Add SplitView e2e tests (#6140)
* Add splitView layout screen to the playground app.
* Add splitView e2e tests.
