+++ e2e/config.json
@@ -2,5 +2,8 @@
   "setupTestFrameworkScriptFile" : "./init.js",
   "testEnvironment": "node",
   "bail": true,
-  "verbose": true
+  "verbose": true,
+  "moduleNameMapper": {
+    ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "identity-obj-proxy"
+  }
 }
