+++ playground/src/testIDs.js
@@ -28,6 +28,8 @@ module.exports = {
   EXTERNAL_COMP_SCREEN_HEADER: 'EXTERNAL_COMPONENT_SCREEN_HEADER',
   TOP_BAR: 'TOP_BAR',
   PUSH_BTN: 'PUSH_BUTTON',
+  PUSH_DETAILS_BTN: 'PUSH_DETAILS_BUTTON',
+  PUSH_MASTER_BTN: 'PUSH_MASTER_BUTTON',
   PUSH_SIDE_MENU_BTN: 'PUSH_SIDE_MENU_BTN',
   SHOW_STATIC_EVENTS_SCREEN: 'SHOW_STATIC_EVENTS_SCREEN',
   POP_BTN: 'POP_BUTTON',
@@ -140,6 +142,7 @@ module.exports = {
   GOTO_BUTTONS_SCREEN: 'GOTO_BUTTONS_SCREEN',
   REGISTER_MODAL_DISMISS_EVENT_BTN: 'REGISTER_MODAL_DISMISS_EVENT_BTN',
   HIDE_PREVIOUS_SCREEN_TOP_BAR: 'HIDE_PREVIOUS_SCREEN_TOP_BAR',
+  COCKTAILS_DETAILS_HEADER: 'COCKTAILS_DETAILS_HEADER',
   
   //External Component Buttons
   EXTERNAL_DISMISS_MODAL_BTN: 'EXTERNAL_DISMISS_MODAL_BTN',
